refactor: rename files to match renamed types
